Update to GetHashCode fix AscensionGameDev#1716 3

+++ Intersect (Core)/GameObjects/LightBase.cs
@@ -68,7 +68,7 @@ namespace Intersect.GameObjects
         /// <returns></returns>
         public override int GetHashCode()
         {
-            return HashCode.Combine(Intensity, (int)Expand, Color, TileX, TileY);
+            return HashCode.Combine(Intensity, Expand, Color, TileX, TileY);
         }
 
         /// <summary>
